+++ src/PSRule/Pipeline/PipelineWriter.cs
@@ -72,17 +72,6 @@ namespace PSRule.Pipeline
             _Writer.WriteWarning(message);
         }
 
-        public void WriteWarning(string message, params object[] args)
-        {
-            if (!ShouldWriteWarning() || string.IsNullOrEmpty(message))
-                return;
-
-            if (args == null || args.Length == 0)
-                WriteWarning(message);
-            else
-                WriteWarning(string.Format(Thread.CurrentThread.CurrentCulture, message, args));
-        }
-
         public virtual bool ShouldWriteWarning()
         {
             return _Writer != null && _Writer.ShouldWriteWarning();